Split route file into aggregates, controller and abstractor directories. Added better id format, shorter string template and removed empty array
